Join threads non-blockingly
